First implementation for dry 1.x compatibility (might need some improvements)
